+++ src/lambda_extract_metrics.py
@@ -1,23 +1,51 @@
 import os
-from typing import Any
 import boto3
 import logging
 from itertools import groupby
 
 from lib.aws import AWSNaming, Boto3ClientCreator, TimestreamTableWriter
+from lib.aws.glue_manager import GlueManager
 from lib.settings import Settings
 from lib.core.constants import SettingConfigs
 
 from lib.metrics_extractor import MetricsExtractorProvider
-from lib.metrics_extractor.metrics_extractor_utils import get_last_update_time
+from lib.metrics_extractor.metrics_extractor_utils import (
+    get_last_update_time,
+    get_earliest_last_update_time_for_resource_set,
+)
 from lib.core.constants import SettingConfigResourceTypes as types
-import json
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
 timestream_client = boto3.client("timestream-write")
 timestream_query_client = boto3.client("timestream-query")
+
+
+def collect_glue_data_quality_result_ids(
+    monitored_environment_name: str,
+    resource_names: list[str],
+    dq_last_update_times: dict,
+    boto3_client_creator: Boto3ClientCreator,
+    aws_client_name: str,
+    timestream_writer: TimestreamTableWriter,
+) -> list[str]:
+    logger.info(
+        f"Collecting Glue Data Quality result IDs at env: {monitored_environment_name}"
+    )
+    min_update_time = get_earliest_last_update_time_for_resource_set(
+        last_update_times=dq_last_update_times,
+        resource_names=resource_names,
+        timestream_writer=timestream_writer,
+    )
+
+    logger.info(f"Extracting Glue Data Quality result IDs since {min_update_time}")
+
+    boto3_client = boto3_client_creator.get_client(aws_client_name=aws_client_name)
+    glue_man = GlueManager(glue_client=boto3_client)
+    result_ids = glue_man.list_data_quality_results(started_after=min_update_time)
+
+    return result_ids
 
 
 def process_individual_resource(
@@ -31,6 +59,7 @@ def process_individual_resource(
     timestream_metrics_table_name: str,
     last_update_times: dict,
     alerts_event_bus_name: str,
+    result_ids: list,
 ):
     logger.info(
         f"Processing: {resource_type}: [{resource_name}] at env:{monitored_environment_name}"
@@ -71,14 +100,18 @@ def process_individual_resource(
         since_time = timestream_writer.get_earliest_writeable_time_for_table()
     logger.info(f"Extracting metrics since {since_time}")
 
-    # # 3. Extract metrics data in form of prepared list of timestream records
+    # # 3. Set Result IDs for Glue Data Quality resources
+    if resource_type == types.GLUE_DATA_QUALITY:
+        metrics_extractor.set_result_ids(result_ids=result_ids)
+
+    # # 4. Extract metrics data in form of prepared list of timestream records
     records, common_attributes = metrics_extractor.prepare_metrics_data(
         since_time=since_time
     )
     metrics_record_count = len(records)
     logger.info(f"Extracted {metrics_record_count} records")
 
-    # # 4. Write extracted data to timestream table
+    # # 5. Write extracted data to timestream table
     metrics_extractor.write_metrics(
         records, common_attributes, timestream_table_writer=timestream_writer
     )
@@ -128,7 +161,19 @@ def process_all_resources_by_env_and_type(
         timestream_write_client=timestream_client,
     )
 
-    # 3. Process each resource of a specific type in a specific environment
+    # 3. Collect Result IDs for all Glue Data Quality resources in a specific environment at once
+    result_ids = []
+    if resource_type == types.GLUE_DATA_QUALITY:
+        result_ids = collect_glue_data_quality_result_ids(
+            monitored_environment_name=monitored_environment_name,
+            resource_names=resource_names,
+            dq_last_update_times=last_update_times.get(resource_type),
+            boto3_client_creator=boto3_client_creator,
+            aws_client_name=aws_client_name,
+            timestream_writer=timestream_man,
+        )
+
+    # 4. Process each resource of a specific type in a specific environment
     for name in resource_names:
         process_individual_resource(
             monitored_environment_name=monitored_environment_name,
@@ -141,6 +186,7 @@ def process_all_resources_by_env_and_type(
             timestream_metrics_table_name=metrics_table_name,
             last_update_times=last_update_times,
             alerts_event_bus_name=alerts_event_bus_name,
+            result_ids=result_ids,
         )