+++ src/lib/metrics_extractor/__init__.py
@@ -3,12 +3,13 @@ from .glue_jobs_metrics_extractor import GlueJobsMetricExtractor
 from .glue_workflows_metrics_extractor import GlueWorkflowsMetricExtractor
 from .glue_crawlers_metrics_extractor import GlueCrawlersMetricExtractor
 from .glue_catalogs_metrics_extractor import GlueCatalogsMetricExtractor
-from .glue_data_quality_metrics_extractor import GlueDataQuailityMetricExtractor
+from .glue_data_quality_metrics_extractor import GlueDataQualityMetricExtractor
 from .lambda_functions_metrics_extractor import LambdaFunctionsMetricExtractor
 from .step_functions_metrics_extractor import StepFunctionsMetricExtractor
 from .metrics_extractor_provider import MetricsExtractorProvider
 from .metrics_extractor_utils import (
     retrieve_last_update_time_for_all_resources,
     get_last_update_time,
+    get_earliest_last_update_time_for_resource_set,
     MetricsExtractorException,
 )