+++ tests/src/lib/digest_service/test_digest_data_extractor.py
@@ -20,6 +20,7 @@ EXPECTED_QUERY_COLUMNS = [
     "execution_time_sec",
     "error_message",
 ]
+EXTRA_DQ_COLUMNS = ["context_type", "glue_table_name", "glue_db_name", "glue_job_name"]
 START_TIME = datetime(2000, 1, 1, 0, 0, 0)
 END_TIME = datetime(2000, 1, 2, 0, 0, 0)
 
@@ -31,6 +32,7 @@ END_TIME = datetime(2000, 1, 2, 0, 0, 0)
         ("scen2", types.GLUE_WORKFLOWS),
         ("scen3", types.LAMBDA_FUNCTIONS),
         ("scen4", types.STEP_FUNCTIONS),
+        ("scen5", types.GLUE_DATA_QUALITY),
         # Not yet implemented for Glue Crawlers and Data Catalogs
     ],
 )
@@ -46,10 +48,13 @@ def test_digest_extractor_get_query(scenario, resource_type):
 
     returned_query = returned_extractor.get_query(START_TIME, END_TIME)
     returned_column_names = re.findall(r"\b(\w+)\b", returned_query)
+
+    expected_columns = EXPECTED_QUERY_COLUMNS
+    if resource_type == types.GLUE_DATA_QUALITY:
+        expected_columns += EXTRA_DQ_COLUMNS
+
     missing_columns = [
-        column
-        for column in EXPECTED_QUERY_COLUMNS
-        if column not in returned_column_names
+        column for column in expected_columns if column not in returned_column_names
     ]
     assert missing_columns == [], f"Missing columns: {missing_columns}"
 
@@ -63,6 +68,7 @@ def test_digest_extractor_get_query(scenario, resource_type):
         ("scen4", types.GLUE_CRAWLERS),
         ("scen5", types.LAMBDA_FUNCTIONS),
         ("scen6", types.STEP_FUNCTIONS),
+        ("scen7", types.GLUE_DATA_QUALITY),
     ],
 )
 @patch("lib.digest_service.digest_data_extractor.TimeStreamQueryRunner")
@@ -99,6 +105,7 @@ def test_digest_extract_runs_with_data(mock_query_runner, scenario, resource_typ
         ("scen4", types.GLUE_CRAWLERS),
         ("scen5", types.LAMBDA_FUNCTIONS),
         ("scen6", types.STEP_FUNCTIONS),
+        ("scen7", types.GLUE_DATA_QUALITY),
     ],
 )
 @patch("lib.digest_service.digest_data_extractor.TimeStreamQueryRunner")
@@ -134,6 +141,7 @@ def test_digest_extract_runs_no_data(mock_query_runner, scenario, resource_type)
         ("scen4", types.GLUE_CRAWLERS),
         ("scen5", types.LAMBDA_FUNCTIONS),
         ("scen6", types.STEP_FUNCTIONS),
+        ("scen7", types.GLUE_DATA_QUALITY),
     ],
 )
 @patch("lib.digest_service.digest_data_extractor.TimeStreamQueryRunner")